% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFPE_dat.R
\docType{data}
\name{FFPE_dat}
\alias{FFPE_dat}
\title{FFPE data on 83 regular genes and 28 patients.}
\format{An object of class \code{"list"}.}
\usage{
data(FFPE_dat)
}
\description{
Data from lung cancer patients.
}
\examples{
data(FFPE_dat)
}
\references{
publication to be added
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
